reorganize disk object storage files
