+++ src/Disks/DiskObjectStorage/MetadataStorages/Web/MetadataStorageFromStaticFilesWebServer.h
@@ -5,11 +5,8 @@
 #include <Disks/DiskObjectStorage/MetadataStorages/MetadataStorageTransactionState.h>
 #include <Disks/DiskObjectStorage/ObjectStorages/Web/WebObjectStorage.h>
 
-
 namespace DB
 {
-
-struct PartitionCommand;
 
 class MetadataStorageFromStaticFilesWebServer final : public IMetadataStorage
 {
